Move all code that deals with files table to db::file
